Add checkpoint folder existence check to prevent FileNotFoundError

+++ train/train_vqgan.py
@@ -57,16 +57,17 @@ def run(cfg: DictConfig):
                 log_folder = folder
         if len(log_folder) > 0:
             ckpt_folder = os.path.join(base_dir, log_folder, 'checkpoints')
-            for fn in os.listdir(ckpt_folder):
-                if fn == 'latest_checkpoint.ckpt':
-                    ckpt_file = 'latest_checkpoint_prev.ckpt'
-                    os.rename(os.path.join(ckpt_folder, fn),
-                              os.path.join(ckpt_folder, ckpt_file))
-            if len(ckpt_file) > 0:
-                cfg.model.resume_from_checkpoint = os.path.join(
-                    ckpt_folder, ckpt_file)
-                print('will start from the recent ckpt %s' %
-                      cfg.model.resume_from_checkpoint)
+            if os.path.exists(ckpt_folder):
+                for fn in os.listdir(ckpt_folder):
+                    if fn == 'latest_checkpoint.ckpt':
+                        ckpt_file = 'latest_checkpoint_prev.ckpt'
+                        os.rename(os.path.join(ckpt_folder, fn),
+                                  os.path.join(ckpt_folder, ckpt_file))
+                if len(ckpt_file) > 0:
+                    cfg.model.resume_from_checkpoint = os.path.join(
+                        ckpt_folder, ckpt_file)
+                    print('will start from the recent ckpt %s' %
+                          cfg.model.resume_from_checkpoint)
 
     # PyTorch Lightning 2.x uses 'devices' and 'accelerator' instead of 'gpus'
     trainer_kwargs = {
